+++ src/i18n/en-actions.ts
@@ -51,8 +51,8 @@ const actionsMessageDescriptors = {
       ${ColonyActions.Recovery} {Recovery}
       ${ColonyActions.EmitDomainReputationPenalty} {Smite}
       ${ColonyActions.EmitDomainReputationReward} {Award}
-      other {Generic}
       ${ColonyActions.Decision} {Decision}
+      other {Generic}
     }`,
 };
