--- src/modules/dashboard/components/ColonyDecisionPreview/DecisionPreviewForm.css
@@ -1,84 +1,0 @@
-@value wideButton: 160px;
-
-.main {
-  padding: 90px 0;
-  height: 100%;
-  width: 100%;
-  background-color: white;
-}
-
-.containerMargin {
-  /*
-   * @NOTE On specs
-   *
-   * The specs designed this for 1440. On this resolution it looks exactly as spec'd.
-   * But as you widen the view port more and more the offset will blur out more and more
-   * and it will start to look centered (due to the percentage margins).
-   *
-   * This is not a bug, this is by design.
-   */
-  margin: 0 13.19% 0 24.3%;
-  width: 900px;
-}
-
-.dividerTop {
-  composes: containerMargin;
-  height: 1px;
-  background-color: color-mod(var(--temp-grey-blue-7) alpha(15%));
-}
-
-.contentContainer {
-  composes: containerMargin;
-  display: flex;
-  align-items: flex-start;
-  justify-content: space-between;
-}
-
-.upperContainer {
-  composes: contentContainer;
-  align-items: baseline;
-}
-
-.tagWrapper {
-  margin-bottom: 13px;
-}
-
-.leftContent {
-  margin-top: 17px;
-  padding-bottom: 30px;
-  width: 460px;
-  border-bottom: 1px solid color-mod(var(--temp-grey-blue-7) alpha(15%));
-  word-break: break-word;
-}
-
-.rightContent {
-  margin-top: 17px;
-}
-
-.userinfo {
-  composes: inlineEllipsis from '~styles/text.css';
-  display: flex;
-  align-items: center;
-  font-size: var(--size-normal);
-  font-weight: var(--weight-bold);
-  color: var(--pink);
-}
-
-.userName {
-  margin-left: 5px;
-}
-
-.title {
-  margin-top: 10px;
-  margin-bottom: 10px;
-}
-
-.details {
-  margin: 0;
-  margin-top: 20px;
-}
-
-.buttonContainer {
-  display: flex;
-  justify-content: flex-end;
-}
