--- src/modules/dashboard/components/ColonyDecisionPreview/DecisionPreviewForm.css.d.ts
@@ -1,14 +1,0 @@
-export const wideButton: string;
-export const main: string;
-export const containerMargin: string;
-export const dividerTop: string;
-export const contentContainer: string;
-export const upperContainer: string;
-export const tagWrapper: string;
-export const leftContent: string;
-export const rightContent: string;
-export const userinfo: string;
-export const userName: string;
-export const title: string;
-export const details: string;
-export const buttonContainer: string;
