--- src/modules/dashboard/components/ColonyDecisionPreview/DetailsWidget/DetailsWidget.css
@@ -1,46 +1,0 @@
-.item {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  min-height: 45px;
-  max-width: 100%;
-  border-top: 1px solid color-mod(var(--temp-grey-blue-7) alpha(15%));
-}
-
-.item:last-child {
-  border-bottom: 1px solid color-mod(var(--temp-grey-blue-7) alpha(15%));
-}
-
-.value, .label {
-  font-size: var(--size-smallish);
-  font-weight: var(--weight-bold);
-  letter-spacing: var(--spacing-medium);
-}
-
-.label {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-right: 40px;
-  line-height: 18px;
-  color: color-mod(var(--temp-grey-blue-7) alpha(85%));
-}
-
-.value {
-  composes: inlineEllipsis from '~styles/text.css';
-  display: flex;
-  align-items: center;
-  max-width: 200px;
-  line-height: 18px;
-  color: color-mod(var(--dark) alpha(85%));
-}
-
-.value i {
-  display: inline-block;
-}
-
-.text {
-  display: inline-block;
-  margin-left: 8px;
-}
