--- src/modules/dashboard/components/ColonyDecisionPreview/DetailsWidget/DetailsWidget.css.d.ts
@@ -1,4 +1,0 @@
-export const item: string;
-export const value: string;
-export const label: string;
-export const text: string;
