--- src/modules/dashboard/components/ColonyDecisionPreview/DetailsWidget/UserWidget/UserWidget.css
@@ -1,31 +1,0 @@
-.textContainer {
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-  flex-direction: row;
-  margin-left: 5px;
-  width: 100%;
-  overflow: hidden;
-}
-
-.textContainer > div {
-  max-width: 100%;
-}
-
-.username {
-  composes: inlineEllipsis from '~styles/text.css';
-  margin-top: -3px;
-  margin-right: 8px;
-  font-size: var(--size-normal);
-  font-weight: var(--weight-bold);
-  color: var(--pink);
-}
-
-.address {
-  composes: inlineEllipsis from '~styles/text.css';
-  font-size: var(--size-smallish);
-  font-weight: var(--weight-bold);
-  line-height: 1;
-  color: var(--temp-grey-blue-7);
-  cursor: pointer;
-}
