--- src/modules/dashboard/components/ColonyDecisionPreview/DetailsWidget/UserWidget/UserWidget.css.d.ts
@@ -1,3 +1,0 @@
-export const textContainer: string;
-export const username: string;
-export const address: string;
