+++ src/modules/dashboard/components/DecisionPreview/DecisionPreviewForm.css
