+++ src/modules/dashboard/components/DecisionPreview/index.ts
@@ -1,1 +1,1 @@
-export { default } from './DecisionPreview';
+export { default } from './DecisionPreviewForm';
